Changed get_or_create to update_or_create
Changed get_or_create to update_or_create, this way if an attribute is changed on the IdP (eg. email/first_name/phone number) the user object and profiles if existing are automatically updated. As is now if they change the lookup fails and therefore get_or_create tries to create a user object but incurs in an IntegrityError on the username field.