Changed get_or_create to update_or_create (#17)
* Changed get_or_create to update_or_create

Changed get_or_create to update_or_create, this way if an attribute is changed on the IdP (eg. email/first_name/phone number) the user object and profiles if existing are automatically updated. As is now if they change the lookup fails and therefore get_or_create tries to create a user object but incurs in an IntegrityError on the username field.

* Fixed update_or_create

* Added default value of 'username' to SAML LOOKUP ATTRIBUTE

* Named arguments before kwargs

* Added choice for attribute sync from IdP

It is now possible to choose wether to sync or not the attributes of the user object with those received from the IdP.

Updated README.md with documentation about use of this settings flag.

Commit by FranceX