Add class views, update auth backend, and make a few structural changes
- Moved the PROVIDERS and USERS_MAP list of dicts to just be dicts with
the provider names as the top level keys. This makes it a little easier
to reference providers and the mapping between the users and providers
no longer requires an enumerate and index operation.
- Remove code because of the above update get_provider_index and ultimately
get_provider_config are not needed.
- Added a settings wrapper object that passes through to the underlying
Django settings. This makes referencing a bit easier and centralizes config
settings.
- Added class-based views and moved the old function view to a different file.
These are a little easier to deal with then the single monolith function view and
it lets us do things with the URL path instead of using query parameters. This
avoids the escaped ampersand issue and makes URL handling a little nicer.
- Added a custom path converter that does a lookup against the provider settings
object and returns the provider to the view. Class-based views are able to use that
provider object instead of doing additional provider lookups.
- Shortened the default path to /saml/. Users can still add additional path to the front
if they so desire.
- Added additional settings and logic to handle lowercase encoding for ADFS
- Added settings and logic to allow users to disable Idp-initiated flows in the provider
configurations.
- Added a setting and logic for assertion replay protection. This uses the Django cache backend
to store the assertion ID. Future requests with the same assertion ID will fail within the timeout
period which is based on the NotOnOrAfter value from the assertion.
- Added logic to store the AuthNRequestID value and verify in the auth backend. It would be preferrable
to use the Session for this but recent changes to SAMESITE cookie handling means we would need to
set the session cookie to samesite=None (BAD!) to get it to work. We use a custom signed cookie to store
this value in the users browser and set it to samesite=None.
- Other minor clean up and changes.