Update SAML failure handling
- Remove the failed redirect and instead have the acs view return
and 400 bad request with optional error information if debug=True for
the service provider.
- We initially tried to logout a user if the saml request failed but
since session cookies arent sent on the return POST from the IdP this
has inaccurate behavior. Now if the user is logged in attempts to
authenticate via SAML with a different username or lookup field that
ultimately fails (i.e. user does not exist) then they get a 400 bad
request. Their old session still exists and they can navigate away from
the failure. This seems better than having the appearance of a successful
request but the session doesnt change.