Mass of v1 updates (#38)
* Move app to shorter name and add django app config

* Run coverage clean prior to python version and fix coverage calc

* Add class views, update auth backend, and make a few structural changes

- Moved the PROVIDERS and USERS_MAP list of dicts to just be dicts with
the provider names as the top level keys. This makes it a little easier
to reference providers and the mapping between the users and providers
no longer requires an enumerate and index operation.
- Remove code because of the above update get_provider_index and ultimately
get_provider_config are not needed.
- Added a settings wrapper object that passes through to the underlying
Django settings. This makes referencing a bit easier and centralizes config
settings.
- Added class-based views and moved the old function view to a different file.
These are a little easier to deal with then the single monolith function view and
it lets us do things with the URL path instead of using query parameters. This
avoids the escaped ampersand issue and makes URL handling a little nicer.
- Added a custom path converter that does a lookup against the provider settings
object and returns the provider to the view. Class-based views are able to use that
provider object instead of doing additional provider lookups.
- Shortened the default path to /saml/. Users can still add additional path to the front
if they so desire.
- Added additional settings and logic to handle lowercase encoding for ADFS
- Added settings and logic to allow users to disable Idp-initiated flows in the provider
configurations.
- Added a setting and logic for assertion replay protection. This uses the Django cache backend
to store the assertion ID. Future requests with the same assertion ID will fail within the timeout
period which is based on the NotOnOrAfter value from the assertion.
- Added logic to store the AuthNRequestID value and verify in the auth backend. It would be preferrable
to use the Session for this but recent changes to SAMESITE cookie handling means we would need to
set the session cookie to samesite=None (BAD!) to get it to work. We use a custom signed cookie to store
this value in the users browser and set it to samesite=None.
- Other minor clean up and changes.

* Update tests to match changes

* Update README and setup.cfg

* Minor clean up

* Add Django example

* Apply linting and pre-commit fixes

* Get rid of the functional urls and views and pipenv

- Remove the functional urls/views to make this simpler
- Pipenv doesnt handle local installs properly so fallback to pip

* Add models for settings

* Add logic for model based providers and auto-gen sp entity and urls

- Add logic to handle storing and generating configs from models,
including the attribute mapping.
- Attribute mapping is really frustrating and non-standard in saml,
added a couple updates since python3-saml sends attributes back as
lists all the time, even if there is only a single value.
- Add NameId mapping capability and make it the default lookup map value
if nothing else is provided
- Add capability to override the http host value to support servers behind
layers of proxies
- Auto-generate the sp entity id and urls based on the host and path

* Clean up

* Add version handling for JSONField and a custom sqlite json field

- Add a somewhat hacky custom json field to support sqlite databases
so our testing doesnt get crazy.
- Update the model and migrations so it is supported on django<3.1

* Update workflow

* Workflow attempt 2

* Update user mapping function and error handling in auth backend

- Updated the mapping function to try and simplify the logic a bit
- Updated the user database calls and wrapped things in try/except
to better handle errors
- Remove SAML_ROUTE completely and let users specify the path when they
include the url patterns.

* Remove the slug and use object id instead. Fix the hacky JSON field to be better.

- Removed the slug since it wasnt helping anyone and use the object id instead.
- Update the hacky JSONField implementation for sqlite to better mimic the one
added in Django 3.1 and add a form field that mimics the json form field.

* Add id field to admin form

* Update SAML failure handling

- Remove the failed redirect and instead have the acs view return
and 400 bad request with optional error information if debug=True for
the service provider.
- We initially tried to logout a user if the saml request failed but
since session cookies arent sent on the return POST from the IdP this
has inaccurate behavior. Now if the user is logged in attempts to
authenticate via SAML with a different username or lookup field that
ultimately fails (i.e. user does not exist) then they get a 400 bad
request. Their old session still exists and they can navigate away from
the failure. This seems better than having the appearance of a successful
request but the session doesnt change.

* Raise exceptions to leverage Django default 400 and 500 response handling

- Raise exceptions and subclass SuspiciousOperation to get default
Django 400 and 500 response handling.

* Add cache exmaple

* Fixup redirect and relay state handling

- We default to "/" and put checks in place to avoid a null RelayState
value being sent back to us which can break the redirect